refactor: extract home screen filters into hook
